requirements/base.txt
Django==3.1
django-model-utils==4.0.0  # Набор полезных базовых классов и утилит для Django
python-dotenv==0.19.0
psycopg2-binary==2.9.1